backend/northstar/requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
django-extensions==2.2.1
djangorestframework==3.10.2
drf-yasg==1.17.0
psycopg2-binary==2.8.3
python-dotenv==0.10.3
herepy==0.6.3.3
